# doc-files/kypo-training-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 2.3.6"
  title: "KYPO Training - API Reference"
host: "localhost:8083"
basePath: "/kypo-rest-training/api/v1"
...
schemes:
- "http"
- "https"
paths:
...
  /clusters/training-definitions/{definitionId}/hints:
    get:
      tags:
      - "Clusters"
      summary: "Get hint cluster."
      description: ""
      operationId: "getHintCluster"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Hint cluster found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /clusters/training-definitions/{definitionId}/hints/sse:
    get:
      tags:
      - "Clusters"
      summary: "Get hint cluster SSE."
      description: ""
      operationId: "getHintClusterSSE"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Hint cluster SSE found."
          schema:
            type: "array"
            items:
              type: "number"
              format: "double"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /clusters/training-definitions/{definitionId}/n-dimensional:
    get:
      tags:
      - "Clusters"
      summary: "Get n-dimensional cluster."
      description: ""
      operationId: "getNDimensionalCluster"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "N-dimensional cluster found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /clusters/training-definitions/{definitionId}/n-dimensional/sse:
    get:
      tags:
      - "Clusters"
      summary: "Get n-dimensional cluster SSE."
      description: ""
      operationId: "getNDimensionalClusterSSE"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "N-dimensional cluster SSE found."
          schema:
            type: "array"
            items:
              type: "number"
              format: "double"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /clusters/training-definitions/{definitionId}/wrong-answers:
    get:
      tags:
      - "Clusters"
      summary: "Get wrong answers cluster."
      description: ""
      operationId: "getWrongAnswersCluster"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Wrong answers cluster found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /clusters/training-definitions/{definitionId}/wrong-answers/sse:
    get:
      tags:
      - "Clusters"
      summary: "Get wrong answers SSE."
      description: ""
      operationId: "getWrongAnswersClusterSSE"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "instanceIds"
        in: "query"
        description: "List of training instance IDs"
        required: false
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "numberOfClusters"
        in: "query"
        description: "Number of clusters"
        required: true
        type: "integer"
        format: "int32"
      - name: "levelId"
        in: "query"
        description: "Level id"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Wrong answer cluster SSE found."
          schema:
            type: "array"
            items:
              type: "number"
              format: "double"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  AbstractLevelVisualizationDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of level."
      title:
        type: "string"
        example: "Training Level1"
        description: "Short textual description of the level."
      max_score:
        type: "integer"
        format: "int32"
        example: 20
        description: "The maximum score a participant can achieve during a level."
      level_type:
        type: "string"
        example: "TRAINING"
        description: "Type of the level."
        enum:
        - "INFO_LEVEL"
        - "TRAINING_LEVEL"
        - "ACCESS_LEVEL"
        - "ASSESSMENT_LEVEL"
      estimated_duration:
        type: "integer"
        format: "int64"
        example: 5
        description: "Estimated time taken by the player to resolve the level."
      order:
        type: "integer"
        format: "int32"
        example: 1
        description: "Order of level among levels in training definition."
    description: "Superclass for classes TrainingLevelVisualizationDTO, AccessLevelVisualizationDTO,\
      \ AssessmentLevelVisualizationDTO and InfoLevelVisualizationDTO"
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
...
  VisualizationInfoDTO:
    type: "object"
    properties:
      training_definition_id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training definition."
      training_definition_title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      training_definition_estimated_duration:
        type: "integer"
        format: "int64"
        example: 5
        description: "Estimated time it takes to finish runs created from this definition."
      levels:
        type: "array"
        description: "Information about all levels in training definition."
        items:
          $ref: "#/definitions/AbstractLevelVisualizationDTO"
    description: "A model includes necessary information about training definition\
      \ which is needed for visualizations."